Meta: Rename fast/faster features to "quick" (#4745)
