+++ source/features/quick-comment-edit.tsx
@@ -26,7 +26,7 @@ function init(): void {
 					<button
 						type="button"
 						role="menuitem"
-						className={`timeline-comment-action btn-link js-comment-edit-button rgh-edit-comments-faster-button ${pageDetect.isDiscussion() ? 'js-discussions-comment-edit-button' : ''}`}
+						className={`timeline-comment-action btn-link js-comment-edit-button rgh-quick-comment-edit-button ${pageDetect.isDiscussion() ? 'js-discussions-comment-edit-button' : ''}`}
 						aria-label="Edit comment"
 					>
 						<PencilIcon/>
