+++ source/refined-github.ts
@@ -16,7 +16,6 @@ import './features/sticky-conversation-list-toolbar.css';
 import './features/always-show-branch-delete-buttons.css';
 import './features/easier-pr-sha-copy.css';
 import './features/repo-stats-spacing.css';
-import './features/emphasize-draft-pr-label.css';
 import './features/clean-notifications.css';
 import './features/fix-first-tab-length.css';
 import './features/align-repository-header.css';
@@ -30,6 +29,7 @@ import './features/hide-empty-profile-status.css';
 import './features/align-issue-labels';
 import './features/clean-pinned-issues';
 import './features/clean-dashboard';
+import './features/hide-low-quality-comments';
 import './features/hide-noisy-newsfeed-events';
 import './features/hide-watch-and-fork-count';
 import './features/minimize-upload-bar';
@@ -45,8 +45,8 @@ import './features/one-key-formatting';
 import './features/tab-to-indent';
 import './features/hide-navigation-hover-highlight';
 import './features/selection-in-new-tab';
-import './features/hide-comments-faster';
-import './features/edit-comments-faster';
+import './features/quick-comment-hiding';
+import './features/quick-comment-edit';
 import './features/open-all-notifications';
 import './features/copy-on-y';
 import './features/profile-hotkey';
@@ -62,7 +62,6 @@ import './features/linkify-code';
 import './features/download-folder-button';
 import './features/linkify-branch-references';
 import './features/open-all-conversations';
-import './features/hide-low-quality-comments';
 import './features/navigate-pages-with-arrow-keys';
 import './features/conversation-links-on-repo-lists';
 import './features/global-conversation-list-filters';
@@ -71,7 +70,7 @@ import './features/sort-conversations-by-update-time'; // Must be after global-c
 import './features/pinned-issues-update-time';
 import './features/latest-tag-button';
 import './features/default-branch-button';
-import './features/faster-pr-diff-options';
+import './features/quick-pr-diff-options';
 import './features/ci-link';
 import './features/sort-milestones-by-closest-due-date';
 import './features/star-repo-hotkey';
@@ -116,7 +115,7 @@ import './features/enable-file-links-in-compare-view'; // Must be before raw-fil
 import './features/raw-file-link';
 import './features/tags-dropdown';
 import './features/pr-filters';
-import './features/edit-files-faster';
+import './features/quick-file-edit';
 import './features/update-pr-from-base-branch';
 import './features/hide-disabled-milestone-sorter';
 import './features/tag-changelog-link';
@@ -171,7 +170,7 @@ import './features/stop-redirecting-in-notification-bar';
 import './features/prevent-link-loss';
 import './features/first-published-tag-for-merged-pr';
 import './features/show-associated-branch-prs-on-fork';
-import './features/faster-reviews';
+import './features/quick-review';
 import './features/fork-source-link-same-view';
 import './features/pr-jump-to-first-non-viewed-file';
 import './features/keyboard-navigation';
@@ -180,7 +179,7 @@ import './features/use-first-commit-message-for-new-prs';
 import './features/linkify-user-edit-history-popup';
 import './features/clean-repo-filelist-actions';
 import './features/prevent-duplicate-pr-submission';
-import './features/hide-label-faster';
+import './features/quick-label-hiding';
 import './features/clean-conversation-headers';
 import './features/stop-pjax-loading-with-esc';
 import './features/highlight-deleted-and-added-files-in-diffs';
@@ -212,7 +211,7 @@ import './features/view-last-pr-deployment';
 import './features/global-search-filters';
 import './features/clean-header-search-field';
 import './features/avoid-accidental-submissions';
-import './features/delete-review-comments-faster';
+import './features/quick-review-comment-deletion';
 import './features/no-unnecessary-split-diff-view';
 import './features/list-pr-for-branch';
 import './features/cancel-wiki-edit-button';
